Dade Coordinate_2D fields final, and implemented unimplemented methods.

+++ src/main/java/Instances/Maps/I_Coordinate.java
@@ -4,17 +4,19 @@ public interface I_Coordinate {
 
     /**
      * Returns the euclidean distance to another {@link I_Coordinate coordinate}. Should return 0 iff this.equals(other)
-     * return true.
+     * return true.  If other is null, or is not of the same runtime type as this, returns -1.
      * @param other a {@link I_Coordinate coordinate}.
-     * @return the euclidean distance to another {@link I_Coordinate coordinate}.
+     * @return the euclidean distance to another {@link I_Coordinate coordinate}. If other is null, or is not of the
+     * same runtime type as this, returns -1.
      */
     float euclideanDistance(I_Coordinate other);
 
     /**
      * Returns the manhattan distance to another {@link I_Coordinate coordinate}. Should return 0 iff this.equals(other)
-     * return true.
+     * return true.  If other is null, or is not of the same runtime type as this, returns -1.
      * @param other a {@link I_Coordinate coordinate}.
-     * @return the manhattan distance to another {@link I_Coordinate coordinate}.
+     * @return the manhattan distance to another {@link I_Coordinate coordinate}.  If other is null, or is not of the
+     * same runtime type as this, returns -1.
      */
-    float manhattanDistance(I_Coordinate other);
+    int manhattanDistance(I_Coordinate other);
 }
